Improve Servitor Share to match retail
Require DataPack!

@reported by: sahar

+++ src/main/java/com/l2jserver/gameserver/network/serverpackets/GMViewCharacterInfo.java
@@ -91,7 +91,7 @@ public class GMViewCharacterInfo extends L2GameServerPacket
 		writeD(_activeChar.getInventory().getTalismanSlots()); // CT2.3
 		writeD(_activeChar.getInventory().canEquipCloak() ? 1 : 0); // CT2.3
 		writeD((int) _activeChar.getPAtk(null));
-		writeD(_activeChar.getPAtkSpd());
+		writeD((int) _activeChar.getPAtkSpd());
 		writeD((int) _activeChar.getPDef(null));
 		writeD(_activeChar.getEvasionRate(null));
 		writeD(_activeChar.getAccuracy());
@@ -99,7 +99,7 @@ public class GMViewCharacterInfo extends L2GameServerPacket
 		writeD((int) _activeChar.getMAtk(null, null));
 		
 		writeD(_activeChar.getMAtkSpd());
-		writeD(_activeChar.getPAtkSpd());
+		writeD((int) _activeChar.getPAtkSpd());
 		
 		writeD((int) _activeChar.getMDef(null, null));
