+++ src/main/java/com/l2jserver/gameserver/network/serverpackets/UserInfo.java
@@ -134,7 +134,7 @@ public final class UserInfo extends L2GameServerPacket
 		writeD(_activeChar.getInventory().getTalismanSlots());
 		writeD(_activeChar.getInventory().canEquipCloak() ? 1 : 0);
 		writeD((int) _activeChar.getPAtk(null));
-		writeD(_activeChar.getPAtkSpd());
+		writeD((int) _activeChar.getPAtkSpd());
 		writeD((int) _activeChar.getPDef(null));
 		writeD(_activeChar.getEvasionRate(null));
 		writeD(_activeChar.getAccuracy());
@@ -142,7 +142,7 @@ public final class UserInfo extends L2GameServerPacket
 		writeD((int) _activeChar.getMAtk(null, null));
 		
 		writeD(_activeChar.getMAtkSpd());
-		writeD(_activeChar.getPAtkSpd());
+		writeD((int) _activeChar.getPAtkSpd());
 		
 		writeD((int) _activeChar.getMDef(null, null));
